[#76] Refactor HeadObject logic
